requirements.txt
Django>=4.2,<5.0
django-guardian>=2.4.0
Pillow>=10.0.0
